requirements.txt
aiofiles==24.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.4.0
asttokens==2.4.1
attrs==23.2.0
bidict==0.23.1
bottle==0.12.25
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
clr-loader==0.2.6
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cycler==0.12.1
debugpy==1.8.2
decorator==5.1.1
defusedxml==0.7.1
docutils==0.19
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.109.2
fonttools==4.53.1
fpdf2==2.7.9
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
ifaddr==0.2.0
ipykernel==6.29.5
ipython==8.26.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jupyter_client==8.6.2
jupyter_core==5.7.2
kaleido==0.2.1
kiwisolver==1.4.5
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.7.1
matplotlib-inline==0.1.7
multidict==6.0.5
nest-asyncio==1.6.0
nicegui==2.5.0
numpy==1.23.2
openpyxl==3.1.5
orjson==3.10.6
packaging==24.1
pandas==1.5.3
parso==0.8.4
patsy==0.5.6
pefile==2023.2.7
pillow==10.4.0
platformdirs==4.2.2
plotly==5.23.0
plotly-express==0.4.1
prompt_toolkit==3.0.47
propcache==0.2.0
proxy_tools==0.1.0
pscript==0.7.7
psutil==6.0.0
pure-eval==0.2.2
pycparser==2.22
pydantic==2.8.0
pydantic_core==2.20.0
Pygments==2.18.0
pyinstaller==6.8.0
pyinstaller-hooks-contrib==2024.7
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-multipart==0.0.9
python-socketio==5.11.3
pythonnet==3.0.3
pytz==2024.1
pywebview==5.1
PyYAML==6.0.1
pyzmq==26.0.3
requests==2.32.3
scikit-posthocs==0.9.0
scipy==1.11.3
seaborn==0.11.2
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
stack-data==0.6.3
starlette==0.36.3
statannotations==0.6.0
statsmodels==0.14.2
tenacity==9.0.0
tornado==6.4.1
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
vbuild==0.8.2
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
wsproto==1.2.0
yarl==1.17.1
